refactor!: migrate monetary values to BIGINT cents for financial precision
Adjustments:
- Domain: contractual_service_execution and contractual_execution_state now use Money
- App: handlers and view models updated to use Money
- Infra: Postgres repositories and query services now read/write cents
- UI: Presentation screens updated to format Money.toDouble()
- DB: Migration for BigInt columns and data backfill
- Quality: Updated analysis_options.yaml with exhaustive_cases: true

+++ lib/features/admin/presentation/screens/widgets/_sla_execution_detail_drawer.dart
@@ -61,7 +61,7 @@ class SlaExecutionDetailDrawer extends StatelessWidget {
                     const SizedBox(height: 16),
                     _InfoField(
                       label: 'Valor Contratual',
-                      value: _currencyFormat.format(item.contractualValue),
+                      value: _currencyFormat.format(item.contractualValue.toDouble()),
                     ),
                     _InfoField(
                       label: 'Multiplicador NoShow',
@@ -70,7 +70,7 @@ class SlaExecutionDetailDrawer extends StatelessWidget {
                     if (item.status == ExecutionStatus.noShow)
                       _InfoField(
                         label: 'Penalidade Calculada',
-                        value: _currencyFormat.format(item.calculatedPenalty),
+                        value: _currencyFormat.format(item.calculatedPenalty.toDouble()),
                       ),
                     const Divider(height: 48),
                     const Text(
